move atb dnr stuff to atb.js, tests
