+++ integration-test/config-mv3.json
@@ -6,6 +6,7 @@
         "background/test-fingerprint.js",
         "background/storage.js",
         "background/click-attribution.js",
+        "background/atb.js",
         "content-scripts/gpc.js"
     ]
 }
