+++ shared/js/background/atb.es6.js
@@ -8,13 +8,17 @@ const settings = require('./settings.es6')
 const parseUserAgentString = require('../shared-utils/parse-user-agent-string.es6')
 const load = require('./load.es6')
 const browserWrapper = require('./wrapper.es6')
-const { setOrUpdateATBdnrRule } = require('./declarative-net-request')
+const { ATB_PARAM_RULE_ID } = require('./declarative-net-request')
+const { ATB_PARAM_PRIORITY } = require('@duckduckgo/ddg2dnr/lib/rulePriorities')
+const { generateDNRRule } = require('@duckduckgo/ddg2dnr/lib/utils')
 
 const ATB_ERROR_COHORT = 'v1-1'
 const ATB_FORMAT_RE = /(v\d+-\d(?:[a-z_]{2})?)$/
 
 // list of accepted params in ATB url
 const ACCEPTED_URL_PARAMS = ['natb', 'cp', 'npi']
+
+const manifestVersion = browserWrapper.getManifestVersion()
 
 const ATB = (() => {
     // regex to match ddg urls to add atb params to.
@@ -24,9 +28,6 @@ const ATB = (() => {
     const ddgAtbURL = 'https://duckduckgo.com/atb.js?'
 
     return {
-        getRegExpAboutPage () {
-            return regExpAboutPage
-        },
 
         shouldUpdateSetAtb (request) {
             return matchPage.test(request.url)
@@ -45,7 +46,7 @@ const ATB = (() => {
             if (!atbSetting) {
                 atbSetting = ATB_ERROR_COHORT
                 settings.updateSetting('atb', ATB_ERROR_COHORT)
-                setOrUpdateATBdnrRule(ATB_ERROR_COHORT, regExpAboutPage)
+                ATB.setOrUpdateATBdnrRule(ATB_ERROR_COHORT)
                 errorParam = '&e=1'
             }
 
@@ -58,10 +59,10 @@ const ATB = (() => {
 
                 if (res.data.updateVersion) {
                     settings.updateSetting('atb', res.data.updateVersion)
-                    setOrUpdateATBdnrRule(res.data.updateVersion, regExpAboutPage)
+                    ATB.setOrUpdateATBdnrRule(res.data.updateVersion)
                 } else if (atbSetting === ATB_ERROR_COHORT) {
                     settings.updateSetting('atb', res.data.version)
-                    setOrUpdateATBdnrRule(res.data.version, regExpAboutPage)
+                    ATB.setOrUpdateATBdnrRule(res.data.version)
                 }
             })
         },
@@ -100,7 +101,7 @@ const ATB = (() => {
             const url = ddgAtbURL + randomValue + '&browser=' + parseUserAgentString().browser
             return load.JSONfromExternalFile(url).then((res) => {
                 settings.updateSetting('atb', res.data.version)
-                setOrUpdateATBdnrRule(res.data.version, regExpAboutPage)
+                ATB.setOrUpdateATBdnrRule(res.data.version)
             }, () => {
                 console.log('couldn\'t reach atb.js for initial server call, trying again')
                 numTries += 1
@@ -173,7 +174,7 @@ const ATB = (() => {
 
                     if (atb) {
                         settings.updateSetting('atb', atb)
-                        setOrUpdateATBdnrRule(atb, regExpAboutPage)
+                        ATB.setOrUpdateATBdnrRule(atb)
                     }
 
                     ATB.finalizeATB(params)
@@ -207,6 +208,43 @@ const ATB = (() => {
             return !settings.getSetting('hasSeenPostInstall') &&
                 !domain.match(regExpPostInstall) &&
                 !domain.match(regExpSoftwarePage)
+        },
+        
+        /**
+        * Creates a DNR rule for ATB parameters
+        * @param {string} atb
+        */
+        setOrUpdateATBdnrRule: (atb) => {
+            if (manifestVersion === 2) {
+                return 
+            }
+
+            if (!atb) {
+                return
+            }
+
+            const atbRule = generateDNRRule({
+                id: ATB_PARAM_RULE_ID,
+                priority: ATB_PARAM_PRIORITY,
+                actionType: 'redirect',
+                redirect: {
+                    transform: {
+                        queryTransform: {
+                            addOrReplaceParams: [{ key: 'atb', value: atb }]
+                        }
+                    }
+                },
+                resourceTypes: ['main_frame'],
+                requestDomains: ['duckduckgo.com'],
+                regexFilter: regExpAboutPage.source
+            })
+
+            if (atbRule?.id === ATB_PARAM_RULE_ID) {
+                chrome.declarativeNetRequest.updateDynamicRules({
+                    removeRuleIds: [atbRule.id],
+                    addRules: [atbRule]
+                })
+            }
         },
 
         async getSurveyURL () {
