+++ shared/js/background/declarative-net-request.js
@@ -15,8 +15,7 @@ import {
 } from '@duckduckgo/ddg2dnr/lib/utils'
 import {
     SERVICE_WORKER_INITIATED_ALLOWING_PRIORITY,
-    USER_ALLOWLISTED_PRIORITY,
-    ATB_PARAM_PRIORITY
+    USER_ALLOWLISTED_PRIORITY
 } from '@duckduckgo/ddg2dnr/lib/rulePriorities'
 
 export const SETTING_PREFIX = 'declarative_net_request-'
@@ -471,43 +470,6 @@ export async function ensureServiceWorkerInitiatedRequestException () {
     await chrome.declarativeNetRequest.updateSessionRules({
         removeRuleIds, addRules
     })
-}
-
-/**
- * Creates a DNR rule for ATB parameters
- * @param {string} atb
- * @param {object} regexFilter
- */
-export async function setOrUpdateATBdnrRule (atb, regexFilter) {
-    if (!(atb && regexFilter)) {
-        return
-    }
-
-    // format regex for json and dnr
-    regexFilter = regexFilter.toString().replace(/\\/g, '\\').replace(/^\//, '').replace(/\/$/, '')
-
-    const atbRule = generateDNRRule({
-        id: ATB_PARAM_RULE_ID,
-        priority: ATB_PARAM_PRIORITY,
-        actionType: 'redirect',
-        redirect: {
-            transform: {
-                queryTransform: {
-                    addOrReplaceParams: [{ key: 'atb', value: atb }]
-                }
-            }
-        },
-        resourceTypes: ['main_frame'],
-        requestDomains: ['duckduckgo.com'],
-        regexFilter
-    })
-
-    if (atbRule?.id === ATB_PARAM_RULE_ID) {
-        await chrome.declarativeNetRequest.updateDynamicRules({
-            removeRuleIds: [atbRule.id],
-            addRules: [atbRule]
-        })
-    }
 }
 
 if (browserWrapper.getManifestVersion() === 3) {
